Dos & DDoD prevention algo
A Denial of Service Attack is the deliberate flooding of a network from attackers that effectively cuts off legitimate users from the websites they host. It is a malicious attack, causing problems not only for the organizations being attacked, but for the customers and online visitors who are affected. We will help you understand what to do if this happens to you, and how to prevent them in the first place.